rename backend modules, disabled removing modules from local temporary.
