+++ ivy/functional/backends/jax/__init__.py
@@ -33,7 +33,7 @@ if not ivy.is_local():
 if not ivy.is_local():
     _module_in_memory = sys.modules[__name__]
 else:
-    global_backend_compiler = sys.modules["ivy.utils.backend.backend_compiler"]
+    global_backend_compiler = sys.modules["ivy.utils.backend.compiler"]
     _module_in_memory = global_backend_compiler.IMPORT_CACHE[__name__]
 
 use = ivy.utils.backend.ContextManager(_module_in_memory)
