+++ ivy/utils/backend/__init__.py
@@ -1,2 +1,2 @@
-from . import backend_handler
-from .backend_handler import *
+from . import handler
+from .handler import *
