+++ ivy/utils/backend/compiler.py
@@ -164,7 +164,7 @@ class ImportTransformer(ast.NodeTransformer):
             tree.body.insert(
                 self.insert_index,
                 ast.ImportFrom(
-                    module="ivy.utils.backend.backend_compiler",
+                    module="ivy.utils.backend.compiler",
                     names=[ast.alias(name="_ivy_fromimport")],
                     level=0,
                 ),
@@ -172,7 +172,7 @@ class ImportTransformer(ast.NodeTransformer):
             tree.body.insert(
                 self.insert_index,
                 ast.ImportFrom(
-                    module="ivy.utils.backend.backend_compiler",
+                    module="ivy.utils.backend.compiler",
                     names=[ast.alias(name="_ivy_absolute_import")],
                     level=0,
                 ),
@@ -343,7 +343,7 @@ def _ivy_import_module(name, package=None):
 
 
 # We shouldn't be able to set the backend on a local Ivy
-modules_to_remove = ["backend_handler"]
+#modules_to_remove = ["utils.backend.handler"]
 
 
 def with_backend(backend: str):
@@ -351,7 +351,7 @@ def with_backend(backend: str):
     ivy_pack = _ivy_import_module("ivy")
     ivy_pack._is_local = True
     backend_module = _ivy_import_module(
-        ivy_pack.backend_handler._backend_dict[backend], ivy_pack.__package__
+        ivy_pack.utils.backend.handler._backend_dict[backend], ivy_pack.__package__
     )
     # TODO temporary
     if backend == "numpy":
@@ -359,17 +359,18 @@ def with_backend(backend: str):
     elif backend == "jax":
         ivy_pack.set_global_attr("RNG", ivy_pack.functional.backends.jax.random.RNG)
     # We know for sure that the backend stack is empty, no need to do backend unsetting
-    ivy_pack.backend_handler._set_backend_as_ivy(
+    ivy_pack.utils.backend.handler._set_backend_as_ivy(
         ivy_pack.__dict__.copy(), ivy_pack, backend_module
     )
     # Remove access to specific modules on local Ivy
-    for module in modules_to_remove:
-        for fn in inspect.getmembers(ivy_pack.__dict__[module], inspect.isfunction):
-            if fn[1].__module__ != module:
-                continue
-            if hasattr(ivy_pack, fn[0]):
-                del ivy_pack.__dict__[fn[0]]
-        del ivy_pack.__dict__[module]
+    # TODO this doesn't work properly atm due to not handling nested module name
+    #for module in modules_to_remove:
+    #    for fn in inspect.getmembers(ivy_pack.__dict__[module], inspect.isfunction):
+    #        if fn[1].__module__ != module:
+    #            continue
+    #        if hasattr(ivy_pack, fn[0]):
+    #            del ivy_pack.__dict__[fn[0]]
+    #    del ivy_pack.__dict__[module]
     ivy_pack.backend_stack.append(backend_module)
     sys.meta_path.remove(FINDER)
     _clear_cache()
